requirements.txt
Flask==3.0.0
line-bot-sdk==3.11.0
python-dotenv==1.0.1
gunicorn==21.2.0
